+++ R/dotplot.R
@@ -34,8 +34,9 @@ setMethod("dotplot", signature(object = "gseaResult"),
                         label_format = label_format, ...)
                 
                 if (NES) {
-                    p <- suppressMessages(p + aes_(color=~NES) + 
-                        scale_color_continuous(low="blue", high="red", name = "NES")
+                    p <- suppressMessages(p + aes_(fill=~NES) + 
+                        # scale_fill_continuous(name = "NES") +
+                        set_enrichplot_color(type = "fill", name = "NES")
                     )
                 }
                 return(p)
@@ -105,8 +106,9 @@ setMethod("dotplot", signature(object = "gseaResultList"),
                       label_format = label_format, ...)
               
               if (NES) {
-                  p <- suppressMessages(p + aes_(color=~NES) + 
-                      scale_color_continuous(low="blue", high="red", name = "NES")
+                  p <- suppressMessages(p + aes_(fill=~NES) + 
+                      # scale_fill_continuous(name = "NES") + 
+                      set_enrichplot_color(type = "fill", name = "NES")
                   )
               }
               return(p)
@@ -146,7 +148,6 @@ dotplot.enrichResult <- function(object, x = "geneRatio", color = "p.adjust",
                              showCategory=10, size=NULL, split = NULL,
                              font.size=12, title = "", orderBy="x",
                              label_format = 30, decreasing=TRUE) {
-
     colorBy <- match.arg(color, c("pvalue", "p.adjust", "qvalue"))
     if (x == "geneRatio" || x == "GeneRatio") {
         x <- "GeneRatio"
@@ -196,15 +197,17 @@ dotplot.enrichResult <- function(object, x = "geneRatio", color = "p.adjust",
 
     df$Description <- factor(df$Description,
                           levels=rev(unique(df$Description[idx])))
-    ggplot(df, aes_string(x=x, y="Description", size=size, color=colorBy)) +
+    p <- ggplot(df, aes_string(x=x, y="Description", size=size, fill=colorBy)) +
         geom_point() +
-        scale_color_continuous(low="red", high="blue", name = color,
-            guide=guide_colorbar(reverse=TRUE)) +
+        aes(shape = I(enrichplot_point_shape)) + 
+        # scale_fill_continuous(name = color) +
+        set_enrichplot_color(type = "fill", name = color) + 
         scale_y_discrete(labels = label_func) +
         ylab(NULL) + ggtitle(title) + theme_dose(font.size) +
-        scale_size(range=c(3, 8)) +
-        guides(size  = guide_legend(order = 1),
-               color = guide_colorbar(order = 2))
+        scale_size(range=c(3, 8))
+    
+    class(p) <- c("enrichplotDot", class(p))
+    return(p)
 }
 
 
@@ -224,11 +227,12 @@ dotplot.enrichResult <- function(object, x = "geneRatio", color = "p.adjust",
 ##' @param colorBy variable that used to color enriched terms,
 ##' e.g. 'pvalue', 'p.adjust' or 'qvalue'
 ##' @importFrom ggplot2 facet_grid
+##' @importFrom rlang check_installed  
 dotplot.compareClusterResult <- function(object, x= "Cluster", colorBy="p.adjust",
                                          showCategory=5, by="geneRatio", size="geneRatio",
                                          split=NULL, includeAll=TRUE,
                                          font.size=12, title="", label_format = 30,
-                                         group = FALSE, shape = FALSE, facet=NULL, strip_width=15) {
+                                         group = FALSE, shape = FALSE, facet=NULL, strip_width=15) {                                   
     color <- NULL
     if (is.null(size)) size <- by ## by may deprecated in future release
 
@@ -262,22 +266,22 @@ dotplot.compareClusterResult <- function(object, x= "Cluster", colorBy="p.adjust
           ggnewscale::new_scale_colour()
     }
 
+
     if (shape) {
+        check_installed('ggstar', 'for `dotplot()` with `shape = TRUE`.')
         ggstar <- "ggstar"
         require(ggstar, character.only=TRUE)
         # p <- p + ggsymbol::geom_symbol(aes_string(symbol = "Cluster", fill = colorBy)) +
         p <- p + ggstar::geom_star(aes_string(starshape="Cluster", fill=colorBy)) +
-            scale_fill_continuous(low="red", high="blue", guide=guide_colorbar(reverse=TRUE))
+            set_enrichplot_color(type = "fill")
     }  else {
-        p <- p +  geom_point(aes_string(color = colorBy))
+        p <- p +  geom_point(aes_string(fill = colorBy)) + 
+            aes(shape = I(enrichplot_point_shape))
     }
 
-    p <- p + scale_color_continuous(low="red", high="blue",
-                    guide=guide_colorbar(reverse=TRUE)) +
+    p <- p + set_enrichplot_color(type = "fill") +
         ylab(NULL) + ggtitle(title) + DOSE::theme_dose(font.size) +
-        scale_size_continuous(range=c(3, 8)) +
-        guides(size  = guide_legend(order = 1),
-                color = guide_colorbar(order = 2))
+        scale_size_continuous(range=c(3, 8))
 
 
     if (!is.null(facet)) {
@@ -287,6 +291,9 @@ dotplot.compareClusterResult <- function(object, x= "Cluster", colorBy="p.adjust
                 labeller = ggplot2::label_wrap_gen(strip_width)) +
             theme(strip.text = element_text(size = 14))
     }
+
+    class(p) <- c("enrichplotDot", class(p))
+    
     return(p)
 }